// src/modules/area-estrategia-desarrollo-comercial/analisis_colegios/application/dto/create-analisis-colegio.dto.ts
export class CreateAnalisisColegioDto {
  codigo_modular?: string;
  codigoModular?: string;
  nombre?: string;
  nombre_colegio?: string;
  nombreColegio?: string;
  correo?: string;
  telefono?: string;
  nivel?: string;
  director?: string;
  tipo?: string;
  ugel?: string;
  departamento?: string;
  distrito?: string;
  zona?: string;
  cantidad_alumnos?: number | string;
  cantidadAlumnos?: number | string;
  direccion?: string;
  direccion_ie?: string;
  direccionIe?: string;
}